+++ Source/DTCollectionViewManagerAnomalyHandler.swift
@@ -110,7 +110,7 @@ public enum DTCollectionViewManagerAnomaly: Equatable, CustomStringConvertible, 
 open class DTCollectionViewManagerAnomalyHandler : AnomalyHandler {
     
     /// Default action to perform when anomaly is detected. Prints debugDescription of anomaly by default.
-    open static var defaultAction : (DTCollectionViewManagerAnomaly) -> Void = { print($0.debugDescription) }
+    public static var defaultAction : (DTCollectionViewManagerAnomaly) -> Void = { print($0.debugDescription) }
     
     /// Action to perform when anomaly is detected. Defaults to `defaultAction`.
     open var anomalyAction: (DTCollectionViewManagerAnomaly) -> Void = DTCollectionViewManagerAnomalyHandler.defaultAction