+++ Tests/SupplementaryEventsTestCase.swift
@@ -55,7 +55,7 @@ class SupplementaryEventsTestCase: XCTestCase {
 
         controller.collectionView?.performBatchUpdates(nil, completion: nil)
         
-        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: UICollectionElementKindSectionHeader, at:  indexPath(0, 0))
+        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: DTCollectionViewElementSectionHeader, at:  indexPath(0, 0))
         
         expect((view as? ReactingHeaderFooterView)?.model) == "FooBar"
     }
@@ -71,7 +71,7 @@ class SupplementaryEventsTestCase: XCTestCase {
         
         controller.collectionView?.performBatchUpdates(nil, completion: nil)
         
-        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: UICollectionElementKindSectionFooter, at:  indexPath(0, 0))
+        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: DTCollectionViewElementSectionFooter, at:  indexPath(0, 0))
         
         expect((view as? ReactingHeaderFooterView)?.model) == "FooBar"
     }