+++ Tests/SupplementaryMappingTestCase.swift
@@ -28,7 +28,7 @@ class SupplementaryMappingTestCase: XCTestCase {
         
         controller.collectionView?.performBatchUpdates(nil, completion: nil)
         
-        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: UICollectionElementKindSectionHeader, at:  indexPath(0, 0))
+        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: DTCollectionViewElementSectionHeader, at:  indexPath(0, 0))
         expect(view).to(beAKindOf(NibHeaderFooterView.self))
     }
     
@@ -38,7 +38,7 @@ class SupplementaryMappingTestCase: XCTestCase {
         
         controller.collectionView?.performBatchUpdates(nil, completion: nil)
         
-        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: UICollectionElementKindSectionFooter, at:  indexPath(0, 0))
+        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: DTCollectionViewElementSectionFooter, at:  indexPath(0, 0))
         expect(view).to(beAKindOf(NibHeaderFooterView.self))
     }
     
@@ -68,24 +68,24 @@ class SupplementaryMappingTestCase: XCTestCase {
     
     func testRegisterSupplementaryClassForKind()
     {
-        controller.manager.registerSupplementary(NibHeaderFooterView.self, forKind: UICollectionElementKindSectionFooter)
+        controller.manager.registerSupplementary(NibHeaderFooterView.self, forKind: DTCollectionViewElementSectionFooter)
         verifyFooter()
     }
     
     func testRegisterNibNamedForSupplementaryClass()
     {
-        controller.manager.registerNibNamed("RandomNameHeaderFooterView", forSupplementary: NibHeaderFooterView.self, ofKind: UICollectionElementKindSectionFooter)
+        controller.manager.registerNibNamed("RandomNameHeaderFooterView", forSupplementary: NibHeaderFooterView.self, ofKind: DTCollectionViewElementSectionFooter)
         verifyFooter()
     }
     
     func testRegisterNiblessSupplementaryClass() {
-        controller.manager.registerNiblessSupplementary(NiblessHeaderFooterView.self, forKind: UICollectionElementKindSectionHeader)
+        controller.manager.registerNiblessSupplementary(NiblessHeaderFooterView.self, forKind: DTCollectionViewElementSectionHeader)
         (controller.collectionView?.collectionViewLayout as? UICollectionViewFlowLayout)?.headerReferenceSize = CGSize(width: 320, height: 50)
         controller.manager.memoryStorage.setSectionHeaderModels([1])
         
         controller.collectionView?.performBatchUpdates(nil, completion: nil)
         
-        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: UICollectionElementKindSectionHeader, at:  indexPath(0, 0))
+        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: DTCollectionViewElementSectionHeader, at:  indexPath(0, 0))
         expect(view).to(beAKindOf(NiblessHeaderFooterView.self))
     }
     
@@ -96,7 +96,7 @@ class SupplementaryMappingTestCase: XCTestCase {
         
         controller.collectionView?.performBatchUpdates(nil, completion: nil)
         
-        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: UICollectionElementKindSectionHeader, at:  indexPath(0, 0))
+        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: DTCollectionViewElementSectionHeader, at:  indexPath(0, 0))
         expect(view).to(beAKindOf(NiblessHeaderFooterView.self))
     }
     
@@ -107,7 +107,7 @@ class SupplementaryMappingTestCase: XCTestCase {
         
         controller.collectionView?.performBatchUpdates(nil, completion: nil)
         
-        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: UICollectionElementKindSectionFooter, at:  indexPath(0, 0))
+        let view = controller.manager.collectionDataSource?.collectionView(controller.collectionView!, viewForSupplementaryElementOfKind: DTCollectionViewElementSectionFooter, at:  indexPath(0, 0))
         expect(view).to(beAKindOf(NiblessHeaderFooterView.self))
     }
 }